drive and pose

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mechanisms_Final/Pose.java
@@ -2,11 +2,11 @@ package org.firstinspires.ftc.teamcode.Mechanisms_Final;
 
 public class Pose {
 
-    public int x;
-    public int y;
-    public int orientation;
+    public double x;
+    public double y;
+    public double orientation;
 
-    public Pose(int x, int y, int orientation) {
+    public Pose(double x, double y, double orientation) {
         this.x = x;
         this.y = y;
         this.orientation = orientation;
